moved getting the username from the cookie to a util module because otherwise it would be a duplicate
