+++ src/views/partials/head.ejs
@@ -4,5 +4,6 @@
     <title>XPLIT</title>
 	<link rel="stylesheet" href="/assets/css/bootstrap.min.css">
 	<link rel="stylesheet" href="/assets/css/bootstrap-select.min.css">
-	<link rel="stylesheet" href="/assets/css/custom.css">
+    <link rel="stylesheet" href="/assets/css/custom.css">
+    <script src="/util/CookieUtilBundled.js"></script>
 </head>
